requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pymongo==4.13.2
python-jose[cryptography]==3.5.0
pwdlib[argon2]==0.2.1
pydantic-settings==2.10.1
email-validator==2.2.0
python-dotenv==1.1.1
